rework function completers (contexts) - function doesn't have to have all spark pacakges in classpath for complitaion (streaming, hive)
