+++ mist-lib/src/main/scala/mist/api/jdsl/JJobDefinition.scala
@@ -1,6 +1,6 @@
 package mist.api.jdsl
 
-import mist.api.BaseContextsArgs
+import mist.api.ContextsArgs
 import mist.api.MistExtras
 import mist.api.jdsl.FuncSyntax._
 import org.apache.spark.api.java.JavaSparkContext
@@ -12,7 +12,7 @@ trait JJobDefinition extends WithArgs {
     * Define job execution that use JavaSparkContext for invocation
     */
   def onSparkContext[R](f: Func1[JavaSparkContext, RetVal[R]]): JHandle[R] = {
-    val job = BaseContextsArgs.javaSparkContext.apply(f.toScalaFunc)
+    val job = ContextsArgs.javaSparkContext.apply(f.toScalaFunc)
     new JHandle(job)
   }
 
@@ -20,7 +20,7 @@ trait JJobDefinition extends WithArgs {
     * Define job execution that use JavaStreamingContext for invocation
     */
   def onStreamingContext[R](f: Func1[JavaStreamingContext, RetVal[R]]): JHandle[R] = {
-    val job = BaseContextsArgs.javaStreamingContext.apply(f.toScalaFunc)
+    val job = ContextsArgs.javaStreamingContext.apply(f.toScalaFunc)
     new JHandle(job)
   }
