+++ mist-lib/src/main/scala/mist/api/package.scala
@@ -1,6 +1,6 @@
 package mist
 
-package object api {
+package object api extends Contexts {
 
   type MLogger = io.hydrosphere.mist.api.MLogger
